changed handle_error and index segments
Right now, the 'segment' prop that is passed to `_layout.svelte` will be undefined for both the index page, and for the error handling page. This will change the segments to `['index']` and `['_error']` respectively. Being able to differentiate between the two pages in the `segment` prop will allow the designer greater control over determining how the page is displayed, such as changing the Nav appearance based upon the segment sent.