Docs: improve docs for skipping SSR
The old title "Making a component SSR compatible" didn't seem quite accurate to me. We're not making the component SSR compatible in the sense that it renders on the server. Rather we're just rendering the component on the client

This also cleans up some of the text and makes it clearer that you can use the same technique for any module you wish to render on the client-side only and not just ones that would fail on the server-side