When getTimings is exposed output the info.
getTimings is exposed from rollup when `perf: true` is set in the rollup
config. Previously there was no way to get this information from a
customer's perspective.

With this change, setting perf: true in the rollup config  will now display
this information to the command line when running `sapper dev`